[DNM] Skip over objects which are known dead ends
Step 1 of #1943

This is an experiment in reducing reads by removing reads for instances of classes that don't hold any reference, at the extra cost of less skipping while indexing (probably fine) and a bit more memory per class.

Current impl adds 1 byte per class, but we could maybe pack that to reuse any unused bit, since all we want is a boolean.